+++ src/main/java/org/jabref/gui/BasePanel.java
@@ -94,7 +94,6 @@ import org.jabref.gui.util.component.CheckBoxMessage;
 import org.jabref.gui.worker.AbstractWorker;
 import org.jabref.gui.worker.CallBack;
 import org.jabref.gui.worker.CitationStyleToClipboardWorker;
-import org.jabref.gui.worker.LookupDOIsWorker;
 import org.jabref.gui.worker.MarkEntriesAction;
 import org.jabref.gui.worker.SendAsEMailAction;
 import org.jabref.logic.autocompleter.AutoCompletePreferences;
@@ -676,7 +675,6 @@ public class BasePanel extends JPanel implements ClipboardOwner, FileUpdateListe
         actions.put(Actions.REMOVE_FROM_GROUP, new GroupAddRemoveDialog(this, false, false));
         actions.put(Actions.MOVE_TO_GROUP, new GroupAddRemoveDialog(this, true, true));
 
-        actions.put(Actions.LOOKUP_DOIS, new LookupDOIsWorker(frame));
         actions.put(Actions.DOWNLOAD_FULL_TEXT, new FindFullTextAction(this));
     }
 
@@ -990,23 +988,7 @@ public class BasePanel extends JPanel implements ClipboardOwner, FileUpdateListe
             if (o instanceof BaseAction) {
                 ((BaseAction) o).action();
             } else {
-                // This part uses Spin's features:
-                Runnable wrk = ((AbstractWorker) o).getWorker();
-                // The Worker returned by getWorker() has been wrapped
-                // by Spin.off(), which makes its methods be run in
-                // a different thread from the EDT.
-                CallBack clb = ((AbstractWorker) o).getCallBack();
-
-                ((AbstractWorker) o).init(); // This method runs in this same thread, the EDT.
-                // Useful for initial GUI actions, like printing a message.
-
-                // The CallBack returned by getCallBack() has been wrapped
-                // by Spin.over(), which makes its methods be run on
-                // the EDT.
-                wrk.run(); // Runs the potentially time-consuming action
-                // without freezing the GUI. The magic is that THIS line
-                // of execution will not continue until run() is finished.
-                clb.update(); // Runs the update() method on the EDT.
+                runWorker((AbstractWorker) o);
             }
         } catch (Throwable ex) {
             // If the action has blocked the JabRefFrame before crashing, we need to unblock it.
@@ -1015,6 +997,26 @@ public class BasePanel extends JPanel implements ClipboardOwner, FileUpdateListe
             frame.unblock();
             LOGGER.error("runCommand error: " + ex.getMessage(), ex);
         }
+    }
+
+    public static void runWorker(AbstractWorker worker) throws Exception {
+        // This part uses Spin's features:
+        Runnable wrk = worker.getWorker();
+        // The Worker returned by getWorker() has been wrapped
+        // by Spin.off(), which makes its methods be run in
+        // a different thread from the EDT.
+        CallBack clb = worker.getCallBack();
+
+        worker.init(); // This method runs in this same thread, the EDT.
+        // Useful for initial GUI actions, like printing a message.
+
+        // The CallBack returned by getCallBack() has been wrapped
+        // by Spin.over(), which makes its methods be run on
+        // the EDT.
+        wrk.run(); // Runs the potentially time-consuming action
+        // without freezing the GUI. The magic is that THIS line
+        // of execution will not continue until run() is finished.
+        clb.update(); // Runs the update() method on the EDT.
     }
 
     private boolean saveDatabase(File file, boolean selectedOnly, Charset enc,