+++ src/main/java/org/jabref/gui/actions/Actions.java
@@ -33,7 +33,6 @@ public class Actions {
     public static final String EXPORT_TO_CLIPBOARD = "exportToClipboard";
     public static final String FOCUS_TABLE = "focusTable";
     public static final String FORWARD = "forward";
-    public static final String LOOKUP_DOIS = "lookupDOIs";
     public static final String MAKE_KEY = "makeKey";
     public static final String MANAGE_SELECTORS = "manageSelectors";
     public static final String MARK_ENTRIES = "markEntries";